requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.20.0
peft>=0.7.0
datasets>=2.14.0
huggingface-hub>=0.16.0
pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
sentencepiece>=0.1.99
protobuf>=3.20.0
einops>=0.6.0
flash-attn>=2.3.0
xformers>=0.0.20
bitsandbytes>=0.41.0
deepspeed>=0.10.0 